requirements.txt
streamlit
openai
langchain
langchain-community
langchain_openai
langchain_pinecone
pinecone-client
tiktoken
python-dotenv